Refactor: remove `flip`
Flip is a bad function. It:

- Reads poorly.
- Is hard to understand.

Originally I thought `flip` was a way to keep the library smaller; now,
I think it kept the library from being complete. I've tended over time
to replace uses of flip with well-named functions, or just remove uses
of flip entirely by rewriting them. For example:

- `flip(fold)` became `over`.
- `extend` was `flip(mixin)`, which became deprecated, then was removed.
- `flatten` used to be `flip(concat)`, which was in turn `flip(push)`.

That last one in particular was convoluted.

Let's establish a general rule of thumb: as soon as you improve a piece
of code by rewriting it to not use a function 3 separate times, remove
the function and rewrite all remaining usages.

The demerits to flip don't end there. It also makes stack traces longer
and harder to humanly parse, and it often takes up more characters than
just writing what you _meant_ in the first place. As a good example,
`flip(concat)` is:

- Harder to read than `append`.
- Hard to interpret as meaning the same thing as `append`.
- More characters to type.
- Useless in stack traces: it says `flip`, not `flip(concat)`.

Even if stack traces _did_ say `flip(concat)`, this still has all the
same disadvantages compared to `append` as `flip(concat)` had in the
first place.

So, yeah. Down with flip! Vivé la clarificación!