Bugfix: at_path: split into get_at_path and set_at_path (for now)
at_path, the general-purpose this idea, is kind of contradictory: it
wants to be a fold, part of the time, like update_path; but, part of the
time, it doesn't actually need to be a fold, because it doesn't actually
care about the rest of the object (as is the case with getters).

So, instead, let's split it into get_at_path and set_at_path for now.
get_at_path loses context from the surrounding object, whereas
set_at_path is a fold, must like update_path. (TODO: set_at_path is not
implemented by this commit. Which is still a bugfix, since get_at_path
now works. It's a halfway step.)

get_at_path and set_at_path could be generalized as folds, but at a
significant additional overhead for keeping track of the surrounding
object in the get_at_path case. It's cheaper, simpler, and more
explicit, to separate get_at_path from e.g. update_path.