Release: v0.3.0: minor improvements, removals, bugfixes, tests
Notably, fixes `get.in(object)(value)` and adds tests for all of the
get.in getters (although most of them are still TODOs).

Refactors mutton mutative-altnernate factories for clarity. These are
not (?) exported, so this _should_ be a non-breaking change; however,
they should probably be exported, and usable in other contexts, so that
explicit mutative alternate are easy to construct even in user code.

Also removes some poorly-conceived shorthand from linchpin; namely, Obj
for Object and Fn for Function. Not necessary, and frankly, confusing.
Not to mention it foils the syntax highlighter, which would normally
color those up.

Fixes the build configuration (which was broken by some change from
earlier). Updates some docs.