Refactor/Major: time: use `performance` instead of `process.hrtime`
- Platform-independent: works in browsers
- Uses doubles and millisecond-precision instead of bigints and nanosecs
- Forces `time` to be asynchronous

Asynchronous profiling may be a frustrating limitation when trying to
instrument (poorly isolated) code, but `time` will not be particularly
applicable to that use-case in the first place, seeing as it has always
been designed to take and evaluate a (synchronous) thunk.

It may be desirable to later refactor `time` to accept promise thunks,
and `await` them in an async function.